+++ iACC/Domain/Card/Card.swift
@@ -1,0 +1,11 @@
+//	
+// Copyright © Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+struct Card: Equatable {
+    let id: Int
+    let number: String
+    let holder: String
+}
