+++ iACC/Domain/Friend/Friend.swift
@@ -1,0 +1,11 @@
+//	
+// Copyright © Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+struct Friend: Equatable {
+    let id: UUID
+    let name: String
+    let phone: String
+}
